pyproject.toml
[project]
name = "elo-adaptive-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "flask>=3.1.2",
    "numpy>=2.3.2",
    "pytz>=2025.2",
    "scikit-learn>=1.7.1",
]
